+++ openrtb2/content.go
@@ -24,7 +24,7 @@ type Content struct {
 	//   integer
 	// Description:
 	//   Episode number.
-	Episode int `json:"episode,omitempty"`
+	Episode int64 `json:"episode,omitempty"`
 
 	// Attribute:
 	//   title
